modular refactoring of Materials, Geoetry and Images
