+++ test/Materials/Materials.jl
@@ -1,7 +1,7 @@
 #############################
 # Materials interface tests #
 #############################
-using Test
+using Test: @testset, @test
 using Apolo.Materials
 
 @testset "Materials unitary tests" begin
@@ -43,31 +43,4 @@ using Apolo.Materials
 
     end
 
-    @testset "SVK material" begin
-
-        material_name = :svk_material
-        svk1 = SVK(E, ν, material_name)
-        material(ν) == material(E) == material_name
-        @test svk1[:ν] == ConstitutiveParameter(:ν, 0.3, (-Inf, Inf), material_name)
-        @test value(svk1, :ν) == νval
-        @test extrema(range(svk1, :E)) == 2 .* (Eₘᵢₙ, Eₘₐₓ)
-        @test !ismissing(svk1, :ν)
-        @test ismissing(svk1, :E)
-        setval!(svk1, :E, 2Eₘᵢₙ)
-        @test value(svk1, :E) == value(E)
-        svk_to_test = replace!(
-            svk1,
-            :E => ConstitutiveParameter(:E, 4.0, (12, 200)),
-        )
-        svk_bench = SVK(
-            ConstitutiveParameter(:E, 4.0, (12, 200)),
-            ConstitutiveParameter(:ν, 0.3, (-Inf, Inf)),
-            material_name)
-        @test svk_to_test == svk_bench
-
-        @test label(svk1) == String(material_name) ==
-              string(material(svk_to_test[:E])) ==
-              string(material(svk_to_test[:ν]))
-
-    end
 end
